rework on vcs flag definition
